:recycle: Move to ast_parser for AST function
